pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faceless"
version = "0.1.0"
description = "Agentic faceless TikTok video pipeline for children's content"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "openai>=1.30.0",
    "pytrends>=4.9.2",
    "google-api-python-client>=2.130.0",
    "moviepy>=1.0.3",
    "Pillow>=10.3.0",
    "structlog>=24.2.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.1",
    "python-multipart>=0.0.9",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "numpy>=1.26.0",
    "pyspellchecker>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
line-length = 100
target-version = "py311"
